*.Tests: reference Core refs to fix build
After we applied the workaround for [1], which included the change
of marking the references of DotNetLightning.Core as private assets,
the build of the tests was failing with errors like:
```
error FS0074: The type referenced through 'Secp256k1Net.Secp256k1' is defined in an assembly that is not referenced. You must add a reference to assembly 'Secp256k1.Net'.
```

[1] https://github.com/joemphilips/DotNetLightning/issues/14