Use MinimumDepth from AcceptChannel message
When we are the funder, the counterparty determines the
MinimumDepth that the funding transaction should reach
before funding_locked.

Instead of incorrectly waiting for the limit that has been
configured in ChannelHandshakeConfig (which should be used
only for inbound channels) use the number from accept_channel,
which is now saved in RemoteParams.

Fixes #38