Flexible ChannelReestablish (fixed) (#48)
Before, we could only generate a channel_establish message if we
received one. Obviously, not every node can work like this.
Furthermore, since we weren't really validating the incoming
message anyway, there was no added benefit to requiring it.

This commit makes it possible to generate a channel_reestablish
message in all states, not just when we are waiting for the funding
transaction to confirm. For example, we need to generate the
message in the Normal state very often.

This fixed version properly replaces ApplyChannelReestablished with
CreateChannelReestablish, which fixes one test, and adds a noop
handler in another test. Tests now pass.