Fix dummy funding tx generation (#73)
The test suite's ProvideFundingTx function would randomly (50% of the
time) return an incorrect funding output index. This is because NBitcoin
was shuffling the funding tx's txouts, but the function would always
return an output index of 0.

This fixes the problem by making ProvideFundingTx disable txout
shuffling on the transaction builder and always return an output index
of 1.