Change BlockHeightOffset from 16bit to 32bit
It's true that, as mentioned in the code, OP_CSV only allows 16bit
offsets. However it is possible to have blocks that are more than
UINT16_MAX blocks apart, and lightning's accept_channel message contains
a 32bit block height offset for the minimum_depth field. Because of
this, the AcceptChannel type was mis-using the 32bit BlockHeight type to
represent a relative block height, even though BlockHeight is intended
to represent absolute block height.

For the sake of using types consistently I've changed BlockHeightOffset
to be 32bits and changed the type of AcceptChannel.MinimumDepth to be a
BlockHeightOffset.