Make init message serialization legacy-compatible
Make init message serialization give special treatment to the
var_onion_optin feature flag. This flag (if it is present) will only be
placed in the global_features section of the init message while other
flags are only placed in the local_features section. This makes use
compatible with legacy clients that distinguish between global and local
features since var_onion_optin was the only global feature in the spec
before the spec was changed and the two feature sets were merged.

Some other miscellaneous changes are:

Add a SetFeature method to the FeatureBit type.

Remove the lazily-initialized mutable bytes field from FeatureBit since
this made it possible for FeatureBit.ByteArray to become out of sync
with FeatureBit.BitArray. ByteArray is now computed from the internal
bitArray on demand.

Modify the BitArray.ToByteArray extension method so that it is the
inverse of the BitArray.FromBytes extension method, rather than the two
methods ordering the bits within the bytes differently.
BitArray.ToByteArray will also return the minimal-length array required
to represent the BitArray rather than (possibly) containing 0x00 bytes
at the front (which is a protocol violation if sent over the network as
a feature set). This also means changing how Payment.TaggedField.WriteTo
serializes feature bits to avoid padding being dropped when padding to a
multiple of 5 bits in order to convert to Bech32.

Change the BitArray.{FromBytes,ToByteArray} property test to take into
account that zero bytes can be dropped from the front of the array.

Add a test case which checks that converting a FeatureBit from and to a
byte array correctly trims zero bytes at the start of the array,
preserves the byte order of the array, and preserves the bit order
within the bytes.

FeatureBit.Equals now just converts both FeatureBits to byte arrays and
tests those for equality. Since converting to bytes normalises the
feature bits by removing excess zeros, this change fixes an issue where
two FeatureBits of different lengths were considered unequal when the
extra bits were all zeros.