Fix bug in fee calculation for first closing fee.
Before this commit, the channel was calculating the fee
with its virtual size instead of what it supposed to give
to `FeeRatePerKw.ToFee`, that is a number of weight unit
for the tx which is 4 times larger than virtual size.
This commit fixes the issue by changing the method name
for the `FeeRatePerKw` to those more explicit.