Reverse the hash when creating from preimage.
When user constructs PaymentRequest from PaymentPreimage,
they usually gets it by `PaymentPreimage.Hash` member. But before this commit,
they had to reverse the order of the hash before giving it to PaymentRequest
constructor(as TaggeedFields). This is not elgonomic, and it breaks the
purpose of the member in the first place.
This commit reverses the order.

This commit also changes the behavior of GetRIPEMD160 method.
(Because we were already using `PaymentPreimage.Hash` for BOLT3 tests, and we
had to reverse the order again before putting it into the commitment secript)