Change inner representation of `PaymentHash`
The correct usage of `uint256` is that it holds
inner value as a little-endian, so that it will
shows big-endian encoded string when we print/parse
the hex value. Before this commit, it had holded it in
big-endian. Which had two problems.

1. It does not align well with `PaymentRequest` construction.
2. It will result to reversed string value when we print/parse.

This commit fixes it by always holding uint256 value in a right way.

This commit also changes the defualt behavior of `PaymentHash.ToBytes()`
so that it will dump big-endian encoded bytes (which probably an user
always wants)