Refactor Macaroons and its tests
Previously, user had to call `initMacaroon` before using 3rd party
Caveats, This was a temporary wrokaround before preparing compiler
flag at solution level, but now we have it. We can switch which build
it uses on `Macaroons/` side.
So move initialization logic to CryptoAlgorithm and switch on macaroons
tests in Macaroon side.

We still do not have 3rd-party caveats support for BouncyCastle build.
This is because, BouncyCastle does not support XSalsa20-Poly1305
(a.k.a. `crypto_secret_box` in NaCl).