Prevent possible F#4.0 issue
It seems that these attributes can cause this compilation error in F#4.0:
error FS0927: The kind of the type specified by its attributes does not match the kind implied by its definition

Newer versions of F# (like 4.5 or even newer, like the one being used
by .NETCore to build the binary that is later published in nuget) allow
compiling this code with no issues, but if you reference the generated
assembly later from an old F# compiler, it could generate exceptions at
runtime, e.g.: System.BadFormatImageException (or other types) whose
inner exception could be the following:

System.TypeLoadException : Could not load type of field 'GWallet.Backend.UtxoCoin.Lightning.SerializedChannel:MinSafeDepth@' (6) due to: Expected reference type but got type kind 17

FSharp.Core's Result type is also affected by this so in this commit we
create a replacement for it that is only used in the BouncyCastle build
(which we now rename as 'Portability' build).

This commit can be reverted once it is assumed that all consumers of
DotNetLightning can use at least F# 4.5. To make sure this happens
even in distros where no Microsoft’s binary blobs are consumed (e.g.
Debian and Ubuntu without using Microsoft apt repositories) we have
proposed an upgrade of the F# toolset downstream, which has been
merged and is in the process of being adopted:

https://salsa.debian.org/dotnet-team/fsharp/-/merge_requests/2

Co-authored-by: Andres G. Aragoneses <knocte@gmail.com>
Co-authored-by: Andrew Cann <shum@canndrew.org>