Refactor shutdown scripts
Previously the local shutdown script was stored in two places in the
channel and also provided as an argument when initiating a shutdown.
This made it possible to send invalid shutdown messages if the shutdown
scripts provided by the user at different times were different. Also,
shutdown scripts have to be in certain forms, but the only place that
they were validated was when initiating a shutdown.

There's now a ShutdownScriptPubKey type which wraps a Script and
enforces that the script is a valid shutdown script. This is used
throughout the code for all scripts which are shutdown scripts, which
forces us to check their validity at all required points.

The DefaultFinalScriptPubKey field has been removed from LocalParams so
that we only store the local shutdown script in (at most) one place in
the channel datastructure. This shutdown script is an Option, and can be
None in the case that we didn't specify a shutdown script to give to the
remote peer when creating the channel. As such as we still pass a
shutdown script as an argument when closing a channel but we also check
that it matches the recorded shutdown script that we previously gave the
remote peer in the case that it's Some.