Remove ChannelWaitingForAcceptChannel.FundingTxFeeRatePerKw
It's not clear if this field is supposed to be different from
InitFeeRatePerKw (eg. to support using a different fee rate for the
funding tx and for the channel). But as it is they were both being used
as the fee rate for commmitment specs so it would have been an error to
set them differently. As such I've merged this field with
InitFeeRatePerKw. Another field could be added later to support using a
different fee rate for the funding tx, but it'll need to be implemented
correctly.