Remove ChannelState.Shutdown
This state is redundant. The `Normal` state already has `Option` fields to
indicate whether we have sent and recieved shutdown messages but, as is,
never sets both of them to `Some` at once and instead transitions to the
Shutdown state when the second one would be set to `Some`. We can
instead just remove this state and have two `Some` values be the
indicator that we are shutting down.

This removes a lot of code-duplication between how messages and events
were handled between the two states.