Change type (again) of NegotiatingData.ClosingTxPerformed
This field had type List<Option<ClosingTxPerformed>>, however a None
could only appear as the final element in this list where it was
ignored anyway. The program logic is identical if we simply make this
field a List<ClosingTxPerformed> and never insert a None at the end of
the list.