Move ShortChannelIdOpt out of NormalData into Channel
The ShortChannelId isn't tied to being in a non-closing state and we may
need to access/modify it when we are in a negotiating state (for
instance, if the channel is shutdown before funding gets locked).
So move it up into Channel.

Also remove the Opt suffix from the field name.