Remove ChannelState
Re-duplicate the local/remote ShutdownScriptPubKey fields since shutdown
scripts sent during channel initialization need to be remembered for the
lifetime of the channel but shutdown scripts sent during shutdown should
be forgotten if the connection is broken before shutdown is complete.

Factor the negotiating data out into the Channel type and remove
ChannelState. The last other remaining state (NormalState) was now
pointless since we store the static shutdown scripts in Channel so that
they can be remembered regardless of what state we're in. The
NegotiatingData field of Channel indicates that we are in shutdown if
both peers have shared shutdown messages.