Seperate incoming and outgoing htlcs in CommitmentSpec
There may be both an incoming and an outgoing htlc for any given htlc
id, so we can't store both incoming and outgoing htlcs in the same map
under the same id.

Separating these into two maps also makes the code a bit less strange
since we no longer have to keep filtering the map for the htlc-direction
that we want.