Update message formats for open/accept channel
For more recent versions of the lightning spec open/accept channel
messages now use a tlv stream at the end of the message to indicate the
presence of an upfront shutdown script. This turns out to be
backwards-compatible, in that the on-the-wire format turns out to be
identical.

If option_upfront_shutdown_script was advertised by both parties, then
the shutdown script field/TLV is mandatory (but may be length 0 to
indicate no shutdown script), otherwise the field is optional (but still
allowed to be present). As such, it seems simpler and more bug-proof to
always include the field and to set it to a length-0 script if we don't
have a shutdown script. Hence the changes to the test.