Fix ApplyFundingLocked
An earlier refactor commit accidently removed the applyEvent handler for
the BothFundingLocked event. As a result, when the executeCommand and
applyEvent branches for the ApplyFundingLocked even got concatenated
into an ApplyFundingLocked method, the branch of code which previously
raised the BothFundingLocked event got replaced with a branch of code
that did nothing. This meant that ApplyFundingLocked ended up doing
nothing whenever we had already called ApplyFundingConfirmedOnBC.

This commit fixes the mistake.