RemoteShutdown: give user local shutdown msg when they need it
Pre-refactor DNL was not giving geewallet the shutdown message when it
should when RemoteShutdown was called. This was hacked around in
geewallet's ClosedChannel code by reaching into DNL's internal state to
get the message.

Post-refactor, the message was no longer stored in DNL's internal state
because it wasn't being used, nor was it being given to geewallet
because that wasn't changed.

This commit fixes this by simply handing it to geewallet when it should.