Use .NET SDK version 6.0 in CI.
If we include net6 to TFM, and use .NET sdk 5.x, the tests will not run even if
we specify `--framework=net5`, since .NET CLI first verifies
the fsproj files and aborts if its `TargetFrameworks` contains the
unknown TFM.

So we now use only latest .net version for tests.

Probably specifying multiple TargetVersions in tests
were not a good idea in the first place.