GitHubCI: merge both publish_* workflows into one
A recent PR tweaking the publish workflows had to duplicate a job
because the 'needs' element doesn't seem to work across workflows.
By merging these two workflows into one, but separating them via jobs,
we can fix this problem so that we get less duplicated code. And the
CI code for publishing to Nuget is very similar anyway, so we might
reuse more lines in the near future as well, there was no need to
separate them.

NOTE: Now the 'on' element of the workflow contains both branches and
tags, but each job checks for what it needs via an 'if' check that
contains a "startsWith" clause, explained here:
https://github.community/t/if-not-startswith-mutually-exclusive-steps/141841/2

And the way to combine two 'if' elements is explained here:
https://github.community/t/how-to-write-multi-line-condition-in-if/128477/8