Mark DotNetLightning.Chain as obsolete
The files in this namespace was originally intended
for injecting external dependencies to `ChannelManager`

But since the architecture of `DotNetLightning.Channel` has
changed a lot, and almost no types defined here is used internally,
we should remove it.
This makes `DotNetLightning.Chain` obsolete as a whole.

The `IFeeEstimator` and `ConfirmationTarget` is still used.
So I have moved it to `DotNetLightning.Channel`