remove plugin logger from published package
`PluginLogger` depends on `Microsoft.Extensions.Logging`
But the version of the library might differ in consumer's side,
which will cause nasty transitive dependency issue.

For now I will keep it simple by simply removing from a published package.
And leave it on the `example/` directory.
(this happens to better in terms language-consistency. Since PluginLogger is written in C#)