Fix null value handling bug in ClnRpc
c-lightning RPC does not allow rpc requests
to have `null` value in some parameters,
instead, if user does not want to specify an parameter,
they simply must not have that field.
This commit fixes null value handling in following cases.

* an optional array field when serializing with System.Text.Json
* Other optional fields with Newtonsoft.Json

It also uses Extension method instead of type extension
for `AddDNLJsonConverters()` so that it will be more C# friendly