stop embedding pdb into dll
It seems that specifying both `SymbolPackageFormat` and
`DebugType = embedded` causes an issue when publishing packages to
nuget.org, since former declares that symbols must be in a separate file.
But the latter declares it does not.
Ending up `.snupkg` files instead of `.nupkg` which supposed to contain
symbol file, but it does not. and nuget.org returns 400 when pushing
those packages.

This commit stops embedding symbols into dll (since that is not
recommended in https://github.com/dotnet/sourcelink).
And publish snupkg files with pdb.