Stop using units of measure for representing msat.
When user wants to use `long`/`int64` in their plugins,
the serializer always think that it is msatoshi. and use
a MSatJsonConverter, if user override it, then Client/Plugin
will not be able to serializer msat.
This is because json serializer does not distinguish units of measure.
Stop using units of measure is a straightforwared way to fix this.